Push type-specific operations down
