+++ scratchstack-database/src/ops/iam/group.rs
@@ -16,3 +16,18 @@ pub use {
     get_group::*, list_attached_group_policies::*, list_groups::*, list_groups_for_user::*, remove_user_from_group::*,
     update_group::*,
 };
+
+use crate::constants::iam::ARN_RESOURCE_PREFIX_GROUP;
+
+/// Return an ARN resource string for a group with the given path and name.
+///
+/// The path is expected to start and end with a slash, but this function will trim extra slashes
+/// if needed.
+pub(crate) fn group_arn_resource(path: &str, group_name: &str) -> String {
+    let resource_path = path.trim_matches('/');
+    if resource_path.is_empty() {
+        format!("{ARN_RESOURCE_PREFIX_GROUP}{group_name}")
+    } else {
+        format!("{ARN_RESOURCE_PREFIX_GROUP}{resource_path}/{group_name}")
+    }
+}
