+++ scratchstack-database/src/ops/iam/user.rs
@@ -15,3 +15,18 @@ pub use {
     attach_user_policy::*, create_user::*, delete_user::*, detach_user_policy::*, get_user::*,
     list_attached_user_policies::*, list_user_tags::*, list_users::*, tag_user::*, untag_user::*, update_user::*,
 };
+
+use crate::constants::iam::ARN_RESOURCE_PREFIX_USER;
+
+/// Return an ARN resource string for a user with the given path and name.
+///
+/// The path is expected to start and end with a slash, but this function will trim extra slashes
+/// if needed.
+fn user_arn_resource(path: &str, user_name: &str) -> String {
+    let resource_path = path.trim_matches('/');
+    if resource_path.is_empty() {
+        format!("{ARN_RESOURCE_PREFIX_USER}{user_name}")
+    } else {
+        format!("{ARN_RESOURCE_PREFIX_USER}{resource_path}/{user_name}")
+    }
+}
